持ち物リスト機能をコンポーネントに分割し、アプリの構造を整理しました。App、Form、PackingList、Item、Logo、Statsコンポーネントを新規作成しました。
