--- src/App.js
@@ -1,179 +1,0 @@
-import React, { useState } from "react";
-
-const initialItems = [
-  { id: 1, description: "パスポート", quantity: 2, packed: false },
-  { id: 2, description: "靴下", quantity: 12, packed: true },
-  { id: 3, description: "充電器", quantity: 1, packed: false },
-];
-
-export default function App() {
-  const [items, setItems] = useState(initialItems);
-  const numItems = items.length;
-
-  function handleAddItems(item) {
-    setItems((items) => [...items, item]);
-  }
-
-  function handleDeleteItem(id) {
-    setItems((items) => items.filter((item) => item.id !== id));
-  }
-
-  function handleToggleItem(id) {
-    setItems((items) =>
-      items.map((item) =>
-        item.id === id ? { ...item, packed: !item.packed } : item
-      )
-    );
-  }
-
-  function handleClearList() {
-    if (window.confirm("リストを削除しますか？")) setItems([]);
-  }
-
-  return (
-    <div className="app">
-      <Logo />
-      <Form onAddItems={handleAddItems} />
-      <PackingList
-        items={items}
-        onDeleteItem={handleDeleteItem}
-        onToggleItem={handleToggleItem}
-        onClearList={handleClearList}
-      />
-      <Stats items={items} />
-    </div>
-  );
-}
-
-function Logo() {
-  return <h1>🏝️ Far Away 🧳</h1>;
-}
-
-function Form({ onAddItems }) {
-  const [description, setDescription] = useState("");
-  const [quantity, setQuantity] = useState(1);
-
-  function handleSubmit(e) {
-    e.preventDefault();
-
-    if (!description.trim()) return;
-
-    const newItem = {
-      id: Date.now(),
-      description,
-      quantity,
-      packed: false,
-    };
-    console.log(newItem);
-
-    onAddItems(newItem);
-
-    setDescription("");
-    setQuantity(1);
-  }
-
-  return (
-    <form className="add-form" onSubmit={handleSubmit}>
-      <h3>😍 旅行に持って行くものは？</h3>
-      <select
-        value={quantity}
-        onChange={(e) => setQuantity(Number(e.target.value))}
-      >
-        {Array.from({ length: 10 }, (_, i) => i + 1).map((num) => (
-          <option value={num} key={num}>
-            {num}
-          </option>
-        ))}
-      </select>
-      <input
-        type="text"
-        placeholder="持ち物..."
-        value={description}
-        onChange={(e) => setDescription(e.target.value)}
-      />
-      <button>追加</button>
-    </form>
-  );
-}
-
-function PackingList({ items, onDeleteItem, onToggleItem, onClearList }) {
-  const [sortBy, setSortBy] = useState("input");
-
-  let sortedItems;
-
-  if (sortBy === "input") sortedItems = items;
-
-  if (sortBy === "quantity")
-    // 数量順
-    sortedItems = items.slice().sort((a, b) => a.quantity - b.quantity);
-
-  if (sortBy === "packed")
-    sortedItems = items
-      .slice()
-      .sort((a, b) => Number(a.packed) - Number(b.packed)); // 完了順
-
-  return (
-    <div className="list">
-      <ul>
-        {sortedItems.map((item) => (
-          <Item
-            item={item}
-            onDeleteItem={onDeleteItem}
-            onToggleItem={onToggleItem}
-            key={item.id}
-          />
-        ))}
-      </ul>
-
-      <div className="actions">
-        <select value={sortBy} onChange={(e) => setSortBy(e.target.value)}>
-          <option value="input">入力順</option>
-          <option value="quantity">数量順</option>
-          <option value="packed">完了順</option>
-        </select>
-        <button onClick={() => setSortBy("input")}>入力順</button>
-        <button onClick={() => setSortBy("quantity")}>数量順</button>
-        <button onClick={() => setSortBy("packed")}>完了順</button>
-        <button onClick={onClearList}>リスト削除</button>
-      </div>
-    </div>
-  );
-}
-
-function Item({ item, onDeleteItem, onToggleItem }) {
-  return (
-    <li>
-      <input
-        type="checkbox"
-        value={item.packed}
-        onChange={() => onToggleItem(item.id)}
-      />
-      <span style={item.packed ? { textDecoration: "line-through" } : {}}>
-        {item.quantity} {item.description}
-      </span>
-      <button onClick={() => onDeleteItem(item.id)}>❎</button>
-    </li>
-  );
-}
-
-function Stats({ items }) {
-  if (!items.length)
-    return (
-      <p className="stats">
-        <em>さあ、旅行の準備をしよう！持ち物を追加してね✈️</em>
-      </p>
-    );
-  const numItems = items.length;
-  const numPacked = items.filter((item) => item.packed).length;
-  const percentage = Math.round((numPacked / numItems) * 100);
-
-  return (
-    <footer className="stats">
-      <em>
-        {percentage === 100
-          ? "🎉 すべての持ち物の準備が完了しました！😄"
-          : `🎒 持ち物リストは${numItems}個で、準備完了した持ち物は${numPacked}個 (${percentage}%)です。`}
-      </em>
-    </footer>
-  );
-}
